Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/trial/requirements.txt
Flask==1.1.2
flask_cors
Flask-OpenTracing==1.1.0
gunicorn==20.0.4
jaeger-client==4.3.0
opentelemetry-api==0.12b0
opentelemetry-exporter-jaeger==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-sdk==0.12b0
opentracing==2.4.0
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.1
pylint==2.4.0
requests==2.23.0
